+++ lib/view/settings/index.dart
@@ -3,16 +3,17 @@ import 'package:get/get.dart';
 import '/view/settings/widgets/button.dart';
 import '/view/settings/widgets/header_section.dart';
 import '/view/settings/widgets/login_btn.dart';
-import 'widgets/lang_btn.dart';
+import 'widgets/language_btn.dart';
 
-class Settings extends StatelessWidget {
-  // const Settings({Key? key}) : super(key: key);
-
+class Settings extends StatefulWidget {
   final bool isConnected;
-  Settings({
-    this.isConnected = false,
-  });
+  Settings({this.isConnected = false});
 
+  @override
+  State<Settings> createState() => _SettingsState();
+}
+
+class _SettingsState extends State<Settings> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -27,7 +28,7 @@ class Settings extends StatelessWidget {
                 child: Column(
               children: <Widget>[
                 // If User is Connected to the app
-                if (isConnected) ...[
+                if (widget.isConnected) ...[
                   Column(
                     crossAxisAlignment: CrossAxisAlignment.end,
                     children: [
@@ -61,55 +62,7 @@ class Settings extends StatelessWidget {
                         textDirection: TextDirection.rtl,
                       ),
                       // Button(onPressed: languageChange, title: "اللغة"),
-                      MaterialButton(
-                        onPressed: languageChange,
-                        child: Ink(
-                          padding: const EdgeInsets.all(8.0),
-                          width: double.infinity,
-                          height: 50,
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(30.0),
-                            color: const Color(0xffF5FAFC),
-                          ),
-                          child: Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            crossAxisAlignment: CrossAxisAlignment.center,
-                            children: <Widget>[
-                              Row(
-                                children: const <Widget>[
-                                  Icon(
-                                    Icons.arrow_back_ios_rounded,
-                                    size: 14,
-                                  ),
-                                  SizedBox(width: 30),
-                                  Text(
-                                    "العربية",
-                                    textDirection: TextDirection.rtl,
-                                    style: TextStyle(
-                                      fontWeight: FontWeight.bold,
-                                      fontSize: 14,
-                                      color: Color(0xff007E6A),
-                                    ),
-                                  ),
-                                ],
-                              ),
-                              const Text(
-                                "اللغة",
-                                textDirection: TextDirection.rtl,
-                                style: TextStyle(
-                                  fontWeight: FontWeight.bold,
-                                  fontSize: 18,
-                                ),
-                              ),
-                            ],
-                          ),
-                        ),
-                        padding: const EdgeInsets.symmetric(
-                            vertical: 10, horizontal: 0.0),
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(32.0),
-                        ),
-                      ),
+                      LanguageBtn(),
                       Button(onPressed: shareApp, title: "شارك التطبيق"),
                       Button(onPressed: contactUs, title: "تواصل معنا"),
                       Button(onPressed: aboutUs, title: "عن التطبيق"),
@@ -123,58 +76,6 @@ class Settings extends StatelessWidget {
         )
       ],
     ));
-  }
-
-  languageChange() {
-    Get.bottomSheet(
-      Container(
-          height: 320,
-          decoration: const BoxDecoration(
-            borderRadius: BorderRadius.only(
-                topLeft: Radius.circular(24), topRight: Radius.circular(24)),
-            color: Color(0xffF5FAFC),
-          ),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: <Widget>[
-              // Logo
-              Container(
-                decoration: const BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.all(Radius.circular(10))),
-                child: Padding(
-                  padding: const EdgeInsets.all(20),
-                  child: Container(
-                    height: 55,
-                    width: 55,
-                    decoration: const BoxDecoration(
-                      image: DecorationImage(
-                        image: AssetImage('assets/language.png'),
-                        fit: BoxFit.fill,
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-              //text
-              const Text(
-                "اختر اللغة",
-                style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20),
-              ),
-              //Buttons
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: const <Widget>[
-                  LanguageBtn(title: "English"),
-                  LanguageBtn(title: "العربية"),
-                ],
-              )
-            ],
-          )),
-      barrierColor: const Color(0xff000029).withOpacity(0.3),
-      isDismissible: false,
-      enableDrag: false,
-    );
   }
 
   shareApp() {
